refactor(): index container providers
