+++ .env.example
@@ -11,3 +11,5 @@ AWS_BUCKET_URL=https://rent-api-br.s3.sa-east-1.amazonaws.com
 APP_API_URL=http://localhost:3333
 ## disk must be "s3" or "local"
 disk=s3
+## disk must be "ses" or "local"
+MAIL_PROVIDER=ses
