+++ src/shared/container/providers/DateProvider/index.ts
@@ -1,0 +1,9 @@
+import "dotenv/config";
+import { container } from "tsyringe";
+import { IDateProvider } from "./IDateProvider";
+import { DayjsDateProvider } from "./implementations/DayjsDateProvider";
+
+container.registerSingleton<IDateProvider>(
+    "DayjsDateProvider",
+    DayjsDateProvider
+);
